Architecture Spike
Dispatch pipeline and credentials, api keys and integrations.
This is outdated and soon to be updated.

Commits:
initial dispatch classes
Initial sketch of dispatch pipeline
Added Submission::Record and some sketches on how to structure it
Design sketches for integrations, credentials, mappings, and encoding/decoding in the pipeline.
credentials
Integration structure, with source, destination, credentials and api keys
Fixed relationships for credential and api key
schema and model for Submission::Record
Implemented and spec'd the integration pipeline
added #designsketch tag to source and destination
comment notes