Fix for minification issue with modules as vars. #153
Interesting enough when specifying a controller for $routeProvider,
the docs say it's either a string or a function. Turns out it
it can take an array like module.controller(...)

In the short term this will help, likely if the whole app had this hack
in place there might be issues.