Refactored the seperation between rails and angular routes. #165
This is an interesting one. Instead of redirecting to
`/?goto=xyz`, I've let the rails view render the angular bootstrapping
layout.

This makes the routing simpler, and easier to earmark places that
are now in angular.

We need to test the navigation stuff a bit, as there might be a
few outliers.

We also need to look out for double logging, but I'm fairly confident
this wont be the case after this.