Fix indentation in products show page; Fix loading of Twitter timeline widget
The $timeout solution is used because the ng-view (products/show) is
rendered after the JavaScript is executed, therefore the element where
the widget should appear is not loaded in the DOM and the JavaScript has
no element to work with.
Please note that in the future, the logic around rendering the widget
should be extracted in a Angular Directive and the "link" directive method should be
the one that takes care of removing the old widget and rendering the new
one between (Angular) page loads.